// src/types/tag-type.ts
/**
 * This enum is by prismarine-nbt. However, due to an incorrect
 * configuration of prismarine-nbt's TypeScript types, importing
 * it from there would not work. Instead, we have copied it over
 * to our own project.
 */
export enum TagType {
  Byte = "byte",
  Short = "short",
  Int = "int",
  Long = "long",
  Float = "float",
  Double = "double",
  ByteArray = "byteArray",
  String = "string",
  List = "list",
  Compound = "compound",
  IntArray = "intArray",
  LongArray = "longArray",
}